+++ navigation-drawer/src/main/java/net/xpece/material/navigationdrawer/list/SupportCompactNavigationListFragment.java
@@ -80,7 +80,7 @@ public class SupportCompactNavigationListFragment extends Fragment
     }
 
     @Override
-    public void setSelectedItem(long id) {
+    public void setSelectedItem(int id) {
         mDelegate.setSelectedItem(id);
     }
